moved complex ops to their own src and header
