+++ H/aops.h
@@ -235,42 +235,3 @@ int32_t outRange_i(CSOUND *csound, OUTRANGE *p);
 int32_t outRange(CSOUND *csound, OUTRANGE *p);
 int32_t hw_channels(CSOUND *csound, ASSIGN *p);
 
-typedef struct {
-  OPDS    h;
-  COMPLEXDAT *ans;
-  COMPLEXDAT *a, *b;
-} CXOP;
-
-typedef struct {
-  OPDS    h;
-  MYFLT *ans;
-  COMPLEXDAT *a, *b;
-} CXOP2R;
-
-
-typedef struct {
-  OPDS    h;
-  COMPLEXDAT *ans;
-  MYFLT *a, *b;
-} R2CXOP;
-
-
-int32_t complex_assign(CSOUND *csound, R2CXOP *p);
-int32_t complex_add(CSOUND *csound, CXOP *p);
-int32_t complex_sub(CSOUND *csound, CXOP *p);
-int32_t complex_prod(CSOUND *csound, CXOP *p);
-int32_t complex_div(CSOUND *csound, CXOP *p);
-int32_t complex_conj(CSOUND *csound, CXOP *p);
-int32_t complex_abs(CSOUND *csound, CXOP2R *p);
-int32_t complex_arg(CSOUND *csound, CXOP2R *p);
-int32_t complex_real(CSOUND *csound, CXOP2R *p);
-int32_t complex_imag(CSOUND *csound, CXOP2R *p);
-int32_t complex_init(CSOUND *csound, CXOP *p);
-int32_t real_add_complex(CSOUND *csound, AOP *p);
-int32_t real_sub_complex(CSOUND *csound, AOP *p);
-int32_t real_mul_complex(CSOUND *csound, AOP *p);
-int32_t real_div_complex(CSOUND *csound, AOP *p);
-int32_t complex_add_real(CSOUND *csound, AOP *p);
-int32_t complex_sub_real(CSOUND *csound, AOP *p);  
-int32_t complex_mul_real(CSOUND *csound, AOP *p);
-int32_t complex_div_real(CSOUND *csound, AOP *p);
